+++ src/sets_graph.jl
@@ -1,23 +1,103 @@
 """
-    Circuit(n_nodes::Int)
+    HamiltonianCycleWeightedType
 
-A Hamiltonian circuit. If the vector `x` is constrained within a `Circuit(n)`,
-each `x[i]` denotes the next node in the graph, for `i ∈ [1, n]`. 
+Whether a Hamiltonian cycle has a weight:
 
-The considered graph is an undirected complete graph with `n` nodes.
-
-Also called `cycle` or `atour`.
+* either the constraint also includes a variable for the total weight:
+  `FIXED_WEIGHT_HAMILTONIAN_CYCLE`
+* or the constraint also includes a variable for the total weight and one 
+  variable giving the weight of each vertex: 
+  `VARIABLE_WEIGHT_HAMILTONIAN_CYCLE`
+* or it only includes the next vertices: `UNWEIGHTED_HAMILTONIAN_CYCLE`
 """
-struct Circuit <: MOI.AbstractVectorSet
-    n_nodes::Int
+@enum HamiltonianCycleWeightedType begin
+    FIXED_WEIGHT_HAMILTONIAN_CYCLE
+    VARIABLE_WEIGHT_HAMILTONIAN_CYCLE
+    UNWEIGHTED_HAMILTONIAN_CYCLE
 end
 
-MOI.dimension(set::Circuit) = set.n_nodes
+"""
+    HamiltonianCycle{HWCT, T}(n_nodes::Int)
+
+A Hamiltonian cycle (i.e. a cycle in a graph that visits each vertex once).
+If the vector `x` is constrained within a `Circuit(n)`, each `x[i]` denotes
+the next vertex in the graph, for `i ∈ [1, n]`. 
+
+The considered graph is an undirected complete graph with `n` vertices.
+To model a Hamiltonian circuit in a noncomplete graph, you can add constraints
+on the variables: if the vertex `i` only has edges towards `j` and `k`, then
+`x[i]` should only have the possible values `j` and `k`.
+
+Also called `cycle` or `atour`.
+
+GCC: https://sofdem.github.io/gccat/gccat/Ccircuit.html
+
+## Unweighted cycle
+
+`HamiltonianCycle{UNWEIGHTED_HAMILTONIAN_CYCLE}` considers an unweighted
+Hamiltonian cycle.
+
+`x`-in-`HamiltonianCycle{UNWEIGHTED_HAMILTONIAN_CYCLE}(n)`:
+a Hamiltonian cycle in the complete graph of `n` vertices.
+`x[i]` is the index of the next vertex in the cycle.
+
+## Fixed-weight cycle
+
+`HamiltonianCycle{FIXED_WEIGHT_HAMILTONIAN_CYCLE}` considers an Hamiltonian
+cycle whose weights are fixed. Having an edge in the cycle increases the
+total weight.
+
+`[x, tw]`-in-`HamiltonianCycle{FIXED_WEIGHT_HAMILTONIAN_CYCLE}(n, weights)`:
+`x` is a Hamiltonian cycle in the complete graph of `n` vertices.
+`x[i]` is the index of the next vertex in the cycle. 
+`tw` is the total weight of the cycle, with `weights` being indexed by the 
+vertices: 
+
+``\\mathtt{tw} = \\sum_{i=1}^{n} \\mathtt{weights[i, x[i]]}``
+
+## Variable-weight cycle
+
+`HamiltonianCycle{VARIABLE_WEIGHT_HAMILTONIAN_CYCLE}` considers an Hamiltonian
+cycle whose weights are variable. Having an edge in the cycle increases the
+total weight.
+
+`[x, w, tw]`-in-`HamiltonianCycle{VARIABLE_WEIGHT_HAMILTONIAN_CYCLE}(n)`:
+`x` is a Hamiltonian cycle in the complete graph of `n` vertices.
+`x[i]` is the index of the next vertex in the cycle. 
+`tw` is the total weight of the cycle, with `w` being indexed by the vertices:
+
+``\\mathtt{tw} = \\sum_{i=1}^{n} \\mathtt{w[i, x[i]]}``
+"""
+struct HamiltonianCycle{HCWT, T <: Real} <: MOI.AbstractVectorSet
+    n_nodes::Int
+    weights::AbstractMatrix{T}
+end
+
+MOI.dimension(set::HamiltonianCycle{FIXED_WEIGHT_HAMILTONIAN_CYCLE, T}) where {T <: Real} = set.n_nodes + 1
+MOI.dimension(set::HamiltonianCycle{VARIABLE_WEIGHT_HAMILTONIAN_CYCLE, T}) where {T <: Real} = 2 * set.n_nodes + 1
+MOI.dimension(set::HamiltonianCycle{UNWEIGHTED_HAMILTONIAN_CYCLE, T}) where {T <: Real} = set.n_nodes
 
 """
-    CircuitPath(n_nodes::Int)
+    HamiltonianPathWeightedType
 
-A Hamiltonian circuit. If the vectors `x` and `y` are constrained within a 
+Whether a Hamiltonian path has a weight:
+
+* either the constraint also includes a variable for the total weight:
+  `WEIGHTED_HAMILTONIAN_PATH`
+* or it only includes the next vertices: `UNWEIGHTED_HAMILTONIAN_PATH`
+"""
+@enum HamiltonianPathWeightedType begin
+    WEIGHTED_HAMILTONIAN_PATH
+    UNWEIGHTED_HAMILTONIAN_PATH
+end
+
+"""
+    HamiltonianPath(n_nodes::Int)
+
+A Hamiltonian path. 
+
+
+If the vectors `x` and `y` are constrained within a 
 `CircuitPath(n)`, each `x[i]` denotes the next node in the graph, for 
 `i ∈ [1, n]`. The last `n` variables denote the order in which the nodes are
 visited, i.e. `y[1]` is the first visited node (1 by convention), `y[2]` is 
@@ -25,11 +105,11 @@ the next node in the path, etc.
 
 The considered graph is an undirected complete graph with `n` nodes.
 """
-struct CircuitPath <: MOI.AbstractVectorSet
+struct HamiltonianPath <: MOI.AbstractVectorSet
     n_nodes::Int
 end
 
-MOI.dimension(set::CircuitPath) = 2 * set.n_nodes
+MOI.dimension(set::HamiltonianPath) = set.n_nodes
 
 """
     WeightedCircuit{T <: Real}(n_nodes::Int, cost_matrix::AbstractMatrix{T})